+++ CHANGELOG.md
@@ -15,6 +15,7 @@ CHANGELOG for FlatCAM beta
 - Tool Calculators: allowed comma to be used as decimal separator
 - changed how the import of svg.path module is done in the ParseSVG.py file
 - Tool Isolation - new feature that allow to isolate interiors of polygons (holes in polygons). It is possible that the isolation to be reported as successful (internal limitations) but some interiors to not be isolated. This way the user get to fix the isolation by doing an extra isolation.
+- added mouse events disconnect in the quit_application() method
 
 4.06.2020